requirements.txt
visdom==0.1.8.9
torch==1.7.0
torchvision==0.8.1
tqdm==4.51.0
numpy==1.19.5
lpips==0.1.3
Pillow==8.0.1
pytorch-fid==0.1.1
